requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
openai
chromadb
pypdf
python-docx
python-dotenv
